requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
discord.py==2.6.4
frozenlist==1.7.0
googletrans==4.0.2
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
langdetect==1.0.9
multidict==6.6.4
numpy==2.0.2
propcache==0.3.2
py-cord==2.6.1
pycparser==2.23
PyNaCl==1.5.0
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
